orchestrator/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
SQLAlchemy==2.0.35
python-multipart==0.0.9
reportlab==4.2.2
PyYAML==6.0.2